REF: move range-generation functions to EA mixin classes (#22016)
